refactor: Clean up project structure and remove package directory
Breaking changes:
Removes the redundant package directory and consolidates project files in the root directory.
