--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,26 +1,0 @@
----
-name: 🐛 Bug Report
-about: Create a report to help us improve
-title: 'bug: '
-labels: bug
-assignees: ''
----
-
-## 🐛 Bug Description
-<!-- Clear and concise description of the bug -->
-
-## 📝 Steps to Reproduce
-1.
-2.
-3.
-
-## 🤔 Expected Behavior
-<!-- What should have happened? -->
-
-## 📸 Screenshots
-<!-- If applicable, add screenshots -->
-
-## 💻 Environment
-- OS: [e.g. macOS, Windows]
-- Node Version: [e.g. 16.x]
-- Package Version: [e.g. 1.7.0]
