--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,22 +1,0 @@
----
-name: ✨ Feature Request
-about: Suggest an idea for this project
-title: 'feat: '
-labels: enhancement
-assignees: ''
----
-
-## 🎯 Feature Description
-<!-- Clear description of the feature you'd like -->
-
-## 🤔 Why is this needed?
-<!-- Explain the use case or problem this would solve -->
-
-## 📝 Proposed Solution
-<!-- How do you think this should work? -->
-
-## 🔄 Alternatives Considered
-<!-- Any alternative solutions you've considered -->
-
-## 📸 Mockups
-<!-- If applicable, add mockups or examples -->
