--- .github/pull_request_template.md
@@ -1,21 +1,0 @@
-## 📝 Description
-<!-- What does this PR do? -->
-
-## 🔄 Type of Change
-- [ ] 🚀 New feature (non-breaking change)
-- [ ] 🐛 Bug fix (non-breaking change)
-- [ ] 💥 Breaking change
-- [ ] 📚 Documentation update
-
-## 🧪 Testing
-<!-- How has this been tested? -->
-
-## 📸 Screenshots
-<!-- If applicable, add screenshots -->
-
-## ✅ Checklist
-- [ ] My code follows the style guidelines
-- [ ] I have performed a self-review
-- [ ] I have commented my code where needed
-- [ ] I have updated the documentation
-- [ ] My changes generate no new warnings
