+++ .husky/commit-msg
@@ -2,3 +2,4 @@
 . "$(dirname -- "$0")/_/husky.sh"
 
 npx --no -- commitlint --edit $1
+npx --no -- commitlint --edit "$1"
