+++ bin/mili-release.js
@@ -1,559 +1,174 @@
 #!/usr/bin/env node
 
 import inquirer from 'inquirer';
-import chalk from 'chalk';
-import { execSync } from 'child_process';
 import fs from 'fs';
 import path from 'path';
+import { execSync } from 'child_process';
+import chalk from 'chalk';
+import os from 'os';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
-import os from 'os';
-import yargs from 'yargs';
-import { hideBin } from 'yargs/helpers';
 
-// ES Module equivalents for __dirname and __filename
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-// Configure CLI commands
-const argv = yargs(hideBin(process.argv))
-  .usage('Usage: $0 [command] [options]')
-  .command('$0', 'Initialize a new Shopify theme project')
-  .option('sync', {
-    alias: 's',
-    description: 'Sync workflows and configurations from @package',
-    type: 'boolean'
-  })
-  .option('connect-github', {
-    alias: 'g',
-    description: 'Connect or reconnect to a GitHub repository',
-    type: 'boolean'
-  })
-  .option('list-stores', {
-    alias: 'l',
-    description: 'List all stored store configurations',
-    type: 'boolean'
-  })
-  .option('use-stored', {
-    alias: 'u',
-    description: 'Use a stored store configuration',
-    type: 'boolean'
-  })
-  .option('remove-store', {
-    alias: 'r',
-    description: 'Remove a stored store configuration',
-    type: 'string'
-  })
-  .option('help', {
-    alias: 'h',
-    description: 'Show help',
-    type: 'boolean'
-  })
-  .help('help')
-  .version()
-  .wrap(null)
-  .epilogue('For more information, visit: https://github.com/milistack')
-  .argv;
-
-// Function to display available commands
-function displayCommands() {
-  console.log(chalk.blue('\nMili Theme CLI - Available Commands\n'));
-
-  console.log(chalk.yellow('Basic Commands:'));
-  console.log('  mili-theme                    Initialize a new theme project');
-  console.log('  mili-theme --help, -h         Show this help message');
-  console.log('  mili-theme --version          Show version number\n');
-
-  console.log(chalk.yellow('Workflow Management:'));
-  console.log('  mili-theme --sync, -s         Sync workflows and configurations');
-  console.log('  npm run sync-workflows        Sync using npm script\n');
-
-  console.log(chalk.yellow('Store Management:'));
-  console.log('  mili-theme --list-stores, -l  List stored store configurations');
-  console.log('  mili-theme --use-stored, -u   Use a stored configuration');
-  console.log('  mili-theme --remove-store, -r Remove a stored configuration\n');
-
-  console.log(chalk.yellow('GitHub Management:'));
-  console.log('  mili-theme --connect-github, -g Connect/reconnect to GitHub repository\n');
-
-  console.log(chalk.yellow('Examples:'));
-  console.log('  # Create new theme project');
-  console.log('  mili-theme\n');
-  console.log('  # Sync workflows after package update');
-  console.log('  mili-theme --sync\n');
-  console.log('  # List stored configurations');
-  console.log('  mili-theme --list-stores\n');
-
-  console.log(chalk.yellow('Documentation:'));
-  console.log('  Full documentation: https://github.com/milistack/theme-cli#readme\n');
+// Helper function to read template files
+function readTemplateFile(filename) {
+  const templatePath = path.join(__dirname, '..', 'templates', filename);
+  return fs.readFileSync(templatePath, 'utf8');
 }
 
-// Get the templates directory path
-const getTemplatesDir = () => path.join(dirname(fileURLToPath(import.meta.url)), '../templates');
+// Helper function to create gitignore content
+function createGitignoreContent() {
+  return `# OS generated files
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
 
-// Helper function to get config directory path
-function getConfigPath() {
-  const configDir = path.join(os.homedir(), '.mili-theme');
-  if (!fs.existsSync(configDir)) {
-    fs.mkdirSync(configDir, { recursive: true });
-  }
-  return path.join(configDir, 'config.json');
+# Environment files
+.env
+.env.*
+
+# Node modules and logs
+node_modules/
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+.pnpm-debug.log*
+
+# Editor directories and files
+.idea/
+.vscode/
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
+
+# Theme specific
+config/settings_data.json
+config.yml
+
+# Build output
+dist/
+build/
+*.zip
+
+# Temporary files
+*.log
+*.tmp
+.cache/
+`;
 }
 
-// Helper function to load stored configurations
-function loadConfig() {
-  const configPath = getConfigPath();
-  try {
-    if (fs.existsSync(configPath)) {
-      return JSON.parse(fs.readFileSync(configPath, 'utf8'));
-    }
-  } catch (error) {
-    console.error(chalk.yellow('Warning: Could not load stored configuration'));
-  }
-  return { stores: {} };
-}
-
-// Helper function to save configurations
-function saveConfig(config) {
-  const configPath = getConfigPath();
-  try {
-    fs.writeFileSync(configPath, JSON.stringify(config, null, 2));
-  } catch (error) {
-    console.error(chalk.yellow('Warning: Could not save configuration'));
-  }
-}
-
-// Helper function to get stored theme token
-async function getThemeToken(storeName, autoUseStored = false) {
-  const config = loadConfig();
-  const storeConfig = config.stores?.[storeName];
-
-  console.log(chalk.blue('\nChecking for stored theme token...'));
-
-  // Add more detailed logging for debugging
-  console.log(chalk.dim('Debug - Store config:', JSON.stringify(storeConfig, null, 2)));
-
-  // Improved validation
-  const hasValidToken = storeConfig?.themeToken &&
-                       typeof storeConfig.themeToken === 'string' &&
-                       storeConfig.themeToken.trim().length > 0;
-
-  if (hasValidToken) {
-    console.log(chalk.green(`Found stored token for ${storeName}`));
-
-    if (autoUseStored) {
-      console.log(chalk.green(`Using stored configuration for ${storeName}`));
-      return storeConfig.themeToken;
-    }
-
-    const { useStored } = await inquirer.prompt([{
-      type: 'confirm',
-      name: 'useStored',
-      message: `Use stored theme token for ${storeName}?`,
-      default: true
-    }]);
-
-    if (useStored) {
-      return storeConfig.themeToken;
-    }
-  } else {
-    console.log(chalk.yellow('No valid stored token found'));
-  }
-
-  const { themeToken } = await inquirer.prompt([{
-    type: 'password',
-    name: 'themeToken',
-    message: 'Enter your Shopify CLI theme token:',
-    validate: input => input.length > 0 || 'Theme token is required'
-  }]);
-
-  const { saveToken } = await inquirer.prompt([{
-    type: 'confirm',
-    name: 'saveToken',
-    message: 'Would you like to save this token for future use?',
-    default: true
-  }]);
-
-  if (saveToken) {
-    config.stores[storeName] = {
-      ...config.stores[storeName],
-      themeToken
-    };
-    saveConfig(config);
-    console.log(chalk.green('Token saved successfully'));
-  }
-
-  return themeToken;
-}
-
-// Helper function to get stored store configuration
-async function getStoredStoreConfig() {
-  const config = loadConfig();
-  const stores = Object.keys(config.stores);
-
-  if (stores.length === 0) {
-    return null;
-  }
-
-  if (stores.length === 1) {
-    const storeName = stores[0];
-    return {
-      storeName,
-      themeToken: config.stores[storeName].themeToken
-    };
-  }
-
-  const { selectedStore } = await inquirer.prompt([{
-    type: 'list',
-    name: 'selectedStore',
-    message: 'Select a stored configuration:',
-    choices: [
-      ...stores.map(store => ({
-        name: store,
-        value: store
-      })),
-      { name: 'Create new configuration', value: null }
-    ]
-  }]);
-
-  if (!selectedStore) {
-    return null;
-  }
-
-  return {
-    storeName: selectedStore,
-    themeToken: config.stores[selectedStore].themeToken
-  };
-}
-
-// Helper function to list stored configurations
-async function listStoredConfigs() {
-  const config = loadConfig();
-  const stores = Object.keys(config.stores);
-
-  if (stores.length === 0) {
-    console.log(chalk.yellow('No stored configurations found.'));
-    return;
-  }
-
-  console.log(chalk.blue('\nStored configurations:'));
-  stores.forEach(store => {
-    console.log(chalk.green(`- ${store}`));
-  });
-}
-
-// Helper function to remove stored configuration
-async function removeStoredConfig(storeName) {
-  const config = loadConfig();
-  if (config.stores[storeName]) {
-    delete config.stores[storeName];
-    saveConfig(config);
-    console.log(chalk.green(`\nRemoved configuration for ${storeName}`));
-  }
-}
-
-async function setupThemes(storeUrl, themeToken, clientName) {
-  console.log(chalk.blue('\nSetting up Shopify themes...\n'));
-
-  // Log masked token for debugging
-  const maskedToken = themeToken ? `${themeToken.substring(0, 4)}...${themeToken.substring(themeToken.length - 4)}` : 'undefined';
-  console.log(chalk.blue('Debug - Theme token:'), chalk.dim(maskedToken));
-  console.log(chalk.blue('Debug - Store URL:'), chalk.dim(storeUrl));
-
-  // Set environment variables for Shopify CLI
-  process.env.SHOPIFY_CLI_THEME_TOKEN = themeToken;
-  process.env.SHOPIFY_FLAG_STORE = storeUrl;
-
-  // Log environment variables for debugging
-  console.log(chalk.blue('Debug - Environment variables set:'));
-  console.log(chalk.dim('SHOPIFY_CLI_THEME_TOKEN:', process.env.SHOPIFY_CLI_THEME_TOKEN ? 'Set' : 'Not set'));
-  console.log(chalk.dim('SHOPIFY_FLAG_STORE:', process.env.SHOPIFY_FLAG_STORE));
-
-  // Helper function to create theme name
-  function createThemeName(type, name) {
-    const fullName = `[${type}] - ${name}`;
-    return fullName.length > 50 ? fullName.substring(0, 47) + '...' : fullName;
-  }
+// Helper function to install GitHub CLI based on OS
+async function installGitHubCLI() {
+  const platform = os.platform();
 
   try {
-    // Verify store access and authentication using theme list instead of store info
-    console.log(chalk.blue('Verifying store access...'));
-    try {
-      // Check if we can access the store by listing themes
-      execSync('shopify theme list --json', { encoding: 'utf8' });
-    } catch (error) {
-      if (error.message.includes('401')) {
-        console.error(chalk.red('\nAuthentication Error: Unable to access the store.'));
-        console.log(chalk.yellow('\nPlease ensure:'));
-        console.log('1. You have logged in to the store directly at least once');
-        console.log(`2. Visit https://${storeUrl}/admin first`);
-        console.log('3. You have the correct permissions (store owner or staff account)');
-        console.log('4. Your theme token is correct');
-        throw new Error('Store authentication failed');
-      }
-      throw error;
-    }
+    console.log(chalk.yellow('\nInstalling GitHub CLI...'));
 
-    // Verify theme permissions
-    console.log(chalk.blue('Verifying theme permissions...'));
-    try {
-      // Try to list themes to verify permissions
-      const themesOutput = execSync('shopify theme list --json', { encoding: 'utf8' });
-      const themes = JSON.parse(themesOutput);
-
-      if (themes.length === 0) {
-        console.log(chalk.yellow('No existing themes found. Creating new themes from scratch.'));
-      } else {
-        // Format themes for selection
-        const themeChoices = themes.map(theme => ({
-          name: `${theme.name}${theme.role === 'live' ? ' (live)' : ''}`,
-          value: theme.id
-        }));
-
-        // Add option to start from scratch
-        themeChoices.push({
-          name: 'Start from scratch (empty theme)',
-          value: null
-        });
-
-        // Prompt for theme selection
-        const { baseTheme } = await inquirer.prompt([
-          {
-            type: 'list',
-            name: 'baseTheme',
-            message: 'Select a theme to use as base:',
-            choices: themeChoices
-          }
-        ]);
-
-        if (baseTheme) {
-          console.log(chalk.blue('\nPulling selected theme as base...'));
-          try {
-            execSync(`shopify theme pull -t ${baseTheme}`, { stdio: 'inherit' });
-          } catch (error) {
-            if (error.message.includes('401') || error.message.includes('403')) {
-              console.error(chalk.red('\nPermission Error: Unable to pull theme.'));
-              console.log(chalk.yellow('\nPlease ensure:'));
-              console.log('1. You have a staff account or are the store owner');
-              console.log('2. Your account has theme management permissions');
-              throw new Error('Theme pull permission denied');
-            }
-            throw error;
-          }
-        }
-      }
-
-      // Check for existing Production and Staging themes
-      const productionTheme = themes.find(t => t.name.startsWith('[Production]'));
-      const stagingTheme = themes.find(t => t.name.startsWith('[Staging]'));
-
-      if (!productionTheme) {
-        const prodThemeName = createThemeName('Production', clientName);
-        console.log(chalk.green(`\nCreating Production theme: ${prodThemeName}...`));
+    switch (platform) {
+      case 'darwin': // macOS
         try {
-          execSync(`shopify theme push --unpublished --json -t "${prodThemeName}"`, { stdio: 'inherit' });
-        } catch (error) {
-          if (error.message.includes('401') || error.message.includes('403')) {
-            console.error(chalk.red('\nPermission Error: Unable to create production theme.'));
-            console.log(chalk.yellow('\nPlease ensure:'));
-            console.log('1. You have a staff account or are the store owner');
-            console.log('2. Your account has theme management permissions');
-            throw new Error('Theme creation permission denied');
-          }
-          throw error;
+          execSync('brew --version', { stdio: 'ignore' });
+        } catch {
+          console.log(chalk.yellow('Installing Homebrew first...'));
+          execSync('/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"', { stdio: 'inherit' });
         }
-      }
+        execSync('brew install gh', { stdio: 'inherit' });
+        break;
 
-      if (!stagingTheme) {
-        const stagingThemeName = createThemeName('Staging', clientName);
-        console.log(chalk.green(`\nCreating Staging theme: ${stagingThemeName}...`));
-        try {
-          execSync(`shopify theme push --unpublished --json -t "${stagingThemeName}"`, { stdio: 'inherit' });
-        } catch (error) {
-          if (error.message.includes('401') || error.message.includes('403')) {
-            console.error(chalk.red('\nPermission Error: Unable to create staging theme.'));
-            console.log(chalk.yellow('\nPlease ensure:'));
-            console.log('1. You have a staff account or are the store owner');
-            console.log('2. Your account has theme management permissions');
-            throw new Error('Theme creation permission denied');
-          }
-          throw error;
+      case 'linux':
+        // Check for different package managers
+        if (fs.existsSync('/etc/debian_version')) {
+          execSync('curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg', { stdio: 'inherit' });
+          execSync('echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null', { stdio: 'inherit' });
+          execSync('sudo apt update && sudo apt install gh -y', { stdio: 'inherit' });
+        } else if (fs.existsSync('/etc/fedora-release')) {
+          execSync('sudo dnf install gh -y', { stdio: 'inherit' });
         }
-      }
-    } catch (error) {
-      if (!error.message.includes('permission denied')) {
-        console.error(chalk.red('\nError verifying theme permissions:'), error.message);
-        console.log(chalk.yellow('\nPlease ensure:'));
-        console.log('1. The Shopify CLI is installed and up to date');
-        console.log('2. You have logged in to the store directly at least once');
-        console.log('3. Your theme token is correct');
-        console.log('4. You have the necessary permissions');
-      }
-      throw error;
+        break;
+
+      case 'win32':
+        execSync('winget install --id GitHub.cli', { stdio: 'inherit' });
+        break;
+
+      default:
+        throw new Error('Unsupported operating system');
     }
 
-    console.log(chalk.green('\n✓ Store access and permissions verified successfully'));
+    console.log(chalk.green('✓ GitHub CLI installed successfully'));
+    return true;
   } catch (error) {
-    console.error(chalk.red('\nError setting up themes:'), error.message);
-    throw error;
+    console.error(chalk.red('Failed to install GitHub CLI:'), error.message);
+    return false;
   }
 }
 
-// Helper function to check if we're in a Git repository
-function isGitRepository() {
+// Helper function to check if gh CLI is installed
+function isGHCliInstalled() {
   try {
-    execSync('git rev-parse --is-inside-work-tree', { stdio: 'ignore' });
+    execSync('gh --version', { stdio: 'ignore' });
     return true;
   } catch (error) {
     return false;
   }
 }
 
-// Helper function to check if we're in the root of a Git repository
-function isGitRoot() {
+// Helper function to check if user is authenticated with gh CLI
+function isGHCliAuthenticated() {
   try {
-    const gitDir = execSync('git rev-parse --git-dir', { encoding: 'utf8' }).trim();
-    return gitDir === '.git';
+    execSync('gh auth status', { stdio: 'ignore' });
+    return true;
   } catch (error) {
     return false;
   }
 }
 
-// Helper function to get the current Git branch
-function getCurrentBranch() {
+// Helper function to authenticate with GitHub CLI
+async function authenticateGitHubCLI() {
   try {
-    return execSync('git rev-parse --abbrev-ref HEAD', { encoding: 'utf8' }).trim();
+    console.log(chalk.yellow('\nAuthenticating with GitHub...'));
+    console.log(chalk.blue('A browser window will open to complete authentication.'));
+
+    execSync('gh auth login -w -p https -s admin:repo', { stdio: 'inherit' });
+
+    console.log(chalk.green('✓ Successfully authenticated with GitHub'));
+    return true;
   } catch (error) {
-    return null;
+    console.error(chalk.red('Failed to authenticate with GitHub:'), error.message);
+    return false;
   }
 }
 
-async function setupGitHub(projectName, shopifyToken, storeUrl) {
-  console.log('\nSetting up GitHub integration...\n');
-
-  const { githubSetup } = await inquirer.prompt([
-    {
-      type: 'list',
-      name: 'githubSetup',
-      message: 'How would you like to set up GitHub?',
-      choices: [
-        { name: 'Create a new repository', value: 'create' },
-        { name: 'Connect to an existing repository', value: 'connect' },
-        { name: 'Skip GitHub setup', value: 'skip' }
-      ]
-    }
-  ]);
-
-  if (githubSetup === 'skip') {
-    console.log('Skipping GitHub setup...');
-    return;
-  }
-
-  // Check if GitHub CLI is installed and authenticated
+// Helper function to create GitHub repository
+async function createGitHubRepo(repoName) {
   try {
-    execSync('gh auth status', { stdio: 'ignore' });
+    console.log(chalk.yellow('\nCreating GitHub repository...'));
+    execSync(`gh repo create ${repoName} --private --confirm`, { stdio: 'inherit' });
+    return true;
   } catch (error) {
-    console.log('\n⚠️  GitHub CLI not installed or not authenticated');
-    console.log('Please install the GitHub CLI and run gh auth login');
-    return;
-  }
-
-  if (githubSetup === 'create') {
-    // Create new repository with project name
-    const repoName = projectName.toLowerCase().replace(/[^a-z0-9-]/g, '-');
-
-    // Ask for confirmation of repository name
-    const { confirmRepoName } = await inquirer.prompt([
-      {
-        type: 'input',
-        name: 'confirmRepoName',
-        message: `Repository will be created as "${repoName}". Press Enter to confirm or type a different name:`,
-        default: repoName
-      }
-    ]);
-
-    console.log(`\nCreating repository ${confirmRepoName}...`);
-
-    try {
-      // First create the repository on GitHub without pushing
-      execSync(`gh repo create "${confirmRepoName}" --private`, { stdio: 'inherit' });
-      console.log(chalk.green('✓ Repository created successfully'));
-
-      // Get GitHub username
-      const username = execSync('gh api user -q .login', { encoding: 'utf8' }).trim();
-
-      // Initialize git if needed
-      if (!isGitRepository()) {
-        execSync('git init', { stdio: 'inherit' });
-      }
-
-      // Create main branch if it doesn't exist
-      try {
-        execSync('git rev-parse --verify main', { stdio: 'ignore' });
-      } catch (error) {
-        execSync('git checkout -b main', { stdio: 'inherit' });
-      }
-
-      // Stage all files
-      execSync('git add .', { stdio: 'inherit' });
-
-      // Create initial commit
-      try {
-        execSync('git commit -m "feat: Initial theme setup"', { stdio: 'inherit' });
-      } catch (error) {
-        if (!error.message.includes('nothing to commit')) {
-          throw error;
-        }
-      }
-
-      // Set up remote
-      try {
-        execSync('git remote remove origin', { stdio: 'pipe' });
-      } catch (error) {
-        // Ignore error if remote doesn't exist
-      }
-
-      execSync(`git remote add origin https://github.com/${username}/${confirmRepoName}.git`, { stdio: 'inherit' });
-      console.log(chalk.green('✓ Remote added successfully'));
-
-      // Push to remote
-      execSync('git push -u origin main --force', { stdio: 'inherit' });
-      console.log(chalk.green('✓ Initial code pushed successfully'));
-
-      // Set up secrets
-      console.log('\nSetting up GitHub secrets...');
-      await setGitHubSecrets(`${username}/${confirmRepoName}`, {
-        SHOPIFY_CLI_THEME_TOKEN: shopifyToken,
-        SHOPIFY_STORE_URL: storeUrl
-      });
-
-    } catch (error) {
-      if (error.message.includes('already exists')) {
-        console.error(chalk.red('Repository already exists. Please choose a different name.'));
-        return;
-      }
-      console.error(chalk.yellow('\nWarning: Error in GitHub setup. You may need to complete these steps manually:'));
-      console.log(chalk.cyan('\n1. Create a repository on GitHub'));
-      console.log(chalk.cyan('2. Run these commands:'));
-      console.log(chalk.cyan('   git init'));
-      console.log(chalk.cyan('   git add .'));
-      console.log(chalk.cyan('   git commit -m "feat: Initial theme setup"'));
-      console.log(chalk.cyan('   git branch -M main'));
-      console.log(chalk.cyan(`   git remote add origin https://github.com/${username}/${confirmRepoName}.git`));
-      console.log(chalk.cyan('   git push -u origin main\n'));
-      console.error(chalk.red(error.message));
-    }
+    console.error(chalk.red('Failed to create repository:'), error.message);
+    return false;
   }
 }
 
 // Helper function to set GitHub secrets
-async function setGitHubSecrets(repoPath, secrets) {
+async function setGitHubSecrets(repoName, secrets) {
   try {
     console.log(chalk.yellow('\nSetting up GitHub secrets...'));
 
+    // Get the username from GitHub CLI
+    const username = execSync('gh api user -q .login', { encoding: 'utf8' }).trim();
+    const repoPath = `${username}/${repoName}`;
+
     for (const [key, value] of Object.entries(secrets)) {
-      // Use gh secret set command with proper flags
-      execSync(`gh secret set ${key} --body "${value}" --repo ${repoPath}`, { stdio: 'inherit' });
+      execSync(`gh secret set ${key} -b"${value}" --repo ${repoPath}`, { stdio: 'inherit' });
       console.log(chalk.green(`✓ Set ${key}`));
     }
 
@@ -564,50 +179,43 @@ async function setGitHubSecrets(repoPath, secrets) {
   }
 }
 
-// Helper function to create or update .gitignore
-function ensureGitignore() {
-  const gitignoreContent = [
-    'node_modules/',
-    '.DS_Store',
-    'config.yml',
-    '.env',
-    '.env.*',
-    '*.log',
-    '.shopifyignore',
-    '.vscode/',
-    '.idea/',
-    'dist/',
-    'build/',
-    'coverage/',
-    '.husky/_/',
-    '.npmrc'
-  ].join('\n');
-
-  const gitignorePath = '.gitignore';
-
+// Helper function to check if repository exists
+async function checkRepoExists(repoName) {
   try {
-    if (fs.existsSync(gitignorePath)) {
-      // Read existing .gitignore
-      const existingContent = fs.readFileSync(gitignorePath, 'utf8');
-      const existingLines = new Set(existingContent.split('\n').map(line => line.trim()));
-
-      // Add new lines that don't exist
-      const newLines = gitignoreContent.split('\n')
-        .filter(line => !existingLines.has(line.trim()));
-
-      if (newLines.length > 0) {
-        // Append new lines to existing .gitignore
-        fs.appendFileSync(gitignorePath, '\n' + newLines.join('\n'));
-        console.log(chalk.green('Updated .gitignore with new entries'));
-      }
-    } else {
-      // Create new .gitignore
-      fs.writeFileSync(gitignorePath, gitignoreContent);
-      console.log(chalk.green('Created new .gitignore file'));
-    }
+    execSync(`gh repo view ${repoName}`, { stdio: 'ignore' });
+    return true;
   } catch (error) {
-    console.error(chalk.yellow('Warning: Could not update .gitignore file'));
-    console.error(error.message);
+    return false;
+  }
+}
+
+// Helper function to check if branch exists
+function branchExists(branchName) {
+  try {
+    execSync(`git show-ref --verify --quiet refs/heads/${branchName}`);
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
+// Helper function to check if branch is protected
+async function isProtectedBranch(repoPath, branchName) {
+  try {
+    const result = execSync(`gh api repos/${repoPath}/branches/${branchName}/protection`, { stdio: 'ignore' });
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
+// Helper function to get remote branch information
+async function getRemoteBranchInfo(repoPath, branchName) {
+  try {
+    const result = execSync(`gh api repos/${repoPath}/branches/${branchName}`, { encoding: 'utf8' });
+    return JSON.parse(result);
+  } catch (error) {
+    return null;
   }
 }
 
@@ -622,7 +230,7 @@ function setupGitHubWorkflows() {
   const workflowTemplatesDir = path.join(templatesDir, 'workflows');
 
   // Copy each workflow file from templates
-  const workflowFiles = ['release.yml', 'preview.yml', 'sync.yml'];
+  const workflowFiles = ['release.yml', 'preview.yml'];
 
   for (const filename of workflowFiles) {
     try {
@@ -633,7 +241,7 @@ function setupGitHubWorkflows() {
         fs.copyFileSync(sourcePath, targetPath);
         console.log(chalk.green(`Created ${filename}`));
       } else {
-        console.error(chalk.yellow(`Warning: Template file ${filename} not found in package templates`));
+        console.error(chalk.yellow(`Warning: Template file ${filename} not found in package templates/workflows`));
       }
     } catch (error) {
       console.error(chalk.yellow(`Warning: Could not create ${filename}`));
@@ -642,32 +250,417 @@ function setupGitHubWorkflows() {
   }
 }
 
+async function init() {
+  console.log(chalk.blue('🚀 Welcome to Mili Release - Shopify Theme Automation'));
+
+  const answers = await inquirer.prompt([
+    {
+      type: 'input',
+      name: 'clientName',
+      message: 'What is your client/project name?',
+      validate: input => input.length > 0
+    },
+    {
+      type: 'input',
+      name: 'storeUrl',
+      message: 'What is your Shopify store URL? (e.g., my-store.myshopify.com)',
+      validate: input => input.includes('.myshopify.com')
+    },
+    {
+      type: 'password',
+      name: 'themeToken',
+      message: 'Enter your Shopify CLI theme token:',
+      validate: input => input.length > 0
+    },
+    {
+      type: 'confirm',
+      name: 'setupGithub',
+      message: 'Would you like to set up GitHub integration?',
+      default: true
+    }
+  ]);
+
+  // Create necessary directories
+  const dirs = ['.github/workflows', '.husky'];
+  dirs.forEach(dir => {
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir, { recursive: true });
+    }
+  });
+
+  // Read and process template files
+  const packageJson = JSON.parse(readTemplateFile('package.json'));
+  packageJson.name = answers.clientName.toLowerCase().replace(/\s+/g, '-');
+
+  // Write files
+  const files = {
+    'package.json': JSON.stringify(packageJson, null, 2),
+    'release.config.js': readTemplateFile('release.config.js'),
+    'commitlint.config.js': readTemplateFile('commitlint.config.js'),
+    '.gitignore': createGitignoreContent()
+  };
+
+  // Ensure .github/workflows directory exists
+  const workflowsDir = '.github/workflows';
+  if (!fs.existsSync(workflowsDir)) {
+    fs.mkdirSync(workflowsDir, { recursive: true });
+  }
+
+  // Copy GitHub Actions workflows
+  setupGitHubWorkflows();
+
+  // Generate README with client details
+  const readmeContent = readTemplateFile('README.md')
+    .replace(/\${clientName}/g, answers.clientName)
+    .replace(/\${storeUrl}/g, answers.storeUrl)
+    .replace(/\${repoName}/g, packageJson.name)
+    .replace(/\${repoUrl}/g, 'https://github.com/USERNAME/REPO.git'); // Will be updated after repo creation
+
+  files['README.md'] = readmeContent;
+
+  // Create theme directory structure
+  const themeDirs = [
+    'assets',
+    'config',
+    'layout',
+    'locales',
+    'sections',
+    'snippets',
+    'templates'
+  ];
+
+  themeDirs.forEach(dir => {
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir, { recursive: true });
+      // Create a .gitkeep file to ensure empty directories are tracked
+      fs.writeFileSync(path.join(dir, '.gitkeep'), '');
+    }
+  });
+
+  Object.entries(files).forEach(([file, content]) => {
+    fs.writeFileSync(file, content);
+  });
+
+  // Create .env file for local development
+  const envContent = `SHOPIFY_FLAG_STORE=${answers.storeUrl}
+SHOPIFY_CLI_THEME_TOKEN=${answers.themeToken}`;
+  fs.writeFileSync('.env', envContent);
+
+  // Initialize git and install dependencies
+  try {
+    execSync('git init', { stdio: 'inherit' });
+    execSync('npm install', { stdio: 'inherit' });
+    execSync('npx husky install', { stdio: 'inherit' });
+    execSync('npx husky add .husky/commit-msg "npx --no -- commitlint --edit $1"', { stdio: 'inherit' });
+
+    if (answers.setupGithub) {
+      let hasGH = isGHCliInstalled();
+      if (!hasGH) {
+        const shouldInstall = await inquirer.prompt([{
+          type: 'confirm',
+          name: 'install',
+          message: 'GitHub CLI is not installed. Would you like to install it?',
+          default: true
+        }]);
+
+        if (shouldInstall.install) {
+          hasGH = await installGitHubCLI();
+        }
+      }
+
+      let isAuthenticated = hasGH && isGHCliAuthenticated();
+      if (hasGH && !isAuthenticated) {
+        const shouldAuth = await inquirer.prompt([{
+          type: 'confirm',
+          name: 'auth',
+          message: 'Would you like to authenticate with GitHub now?',
+          default: true
+        }]);
+
+        if (shouldAuth.auth) {
+          isAuthenticated = await authenticateGitHubCLI();
+        }
+      }
+
+      if (hasGH && isAuthenticated) {
+        // Get GitHub username
+        const username = execSync('gh api user -q .login', { encoding: 'utf8' }).trim();
+
+        // Ask about repository setup
+        const repoSetup = await inquirer.prompt([
+          {
+            type: 'list',
+            name: 'repoAction',
+            message: 'How would you like to set up the GitHub repository?',
+            choices: [
+              { name: 'Create a new repository', value: 'create' },
+              { name: 'Connect to an existing repository', value: 'connect' }
+            ]
+          },
+          {
+            type: 'input',
+            name: 'existingRepo',
+            message: 'Enter the existing repository name (e.g., username/repo):',
+            when: (answers) => answers.repoAction === 'connect',
+            validate: async (input) => {
+              if (!input.includes('/')) {
+                return 'Please enter the repository name in the format username/repo';
+              }
+              const exists = await checkRepoExists(input);
+              return exists ? true : 'Repository not found. Please check the name and try again.';
+            }
+          }
+        ]);
+
+        let repoName = '';
+        let repoPath = '';
+
+        if (repoSetup.repoAction === 'create') {
+          repoName = packageJson.name;
+          repoPath = `${username}/${repoName}`;
+          const repoCreated = await createGitHubRepo(repoName);
+          if (!repoCreated) {
+            throw new Error('Failed to create repository');
+          }
+        } else {
+          repoPath = repoSetup.existingRepo;
+          repoName = repoPath.split('/')[1];
+        }
+
+        // Add remote origin
+        const repoUrl = `https://github.com/${repoPath}.git`;
+        execSync(`git remote add origin ${repoUrl}`, { stdio: 'inherit' });
+
+        // Update README with correct repository URL
+        const readmeContent = fs.readFileSync('README.md', 'utf8')
+          .replace('https://github.com/USERNAME/REPO.git', repoUrl);
+        fs.writeFileSync('README.md', readmeContent);
+
+        // Make initial commit and push
+        execSync('git add .', { stdio: 'inherit' });
+        execSync('git commit -m "feat: Initial theme setup"', { stdio: 'inherit' });
+
+        if (repoSetup.repoAction === 'connect') {
+          // For existing repos, ask which branch to use
+          const branchSetup = await inquirer.prompt([
+            {
+              type: 'list',
+              name: 'branchAction',
+              message: 'How would you like to handle the branch setup?',
+              choices: [
+                { name: 'Create and push to a new branch', value: 'new' },
+                { name: 'Push to an existing branch (Warning: Use with caution)', value: 'existing' }
+              ]
+            },
+            {
+              type: 'input',
+              name: 'branchName',
+              message: 'Enter the branch name:',
+              when: (answers) => answers.branchAction === 'new',
+              default: 'development'
+            },
+            {
+              type: 'input',
+              name: 'existingBranch',
+              message: 'Enter the existing branch name:',
+              when: (answers) => answers.branchAction === 'existing',
+              default: 'development'
+            }
+          ]);
+
+          if (branchSetup.branchAction === 'new') {
+            const branchName = branchSetup.branchName;
+            if (branchExists(branchName)) {
+              console.log(chalk.yellow(`\nBranch '${branchName}' already exists locally. Using existing branch.`));
+              execSync(`git checkout ${branchName}`, { stdio: 'inherit' });
+            } else {
+              execSync(`git checkout -b ${branchName}`, { stdio: 'inherit' });
+            }
+            execSync(`git push -u origin ${branchName}`, { stdio: 'inherit' });
+          } else {
+            const branchName = branchSetup.existingBranch;
+            const isProtected = await isProtectedBranch(repoPath, branchName);
+            const branchInfo = await getRemoteBranchInfo(repoPath, branchName);
+
+            if (isProtected) {
+              console.log(chalk.red(`\n⚠️  WARNING: Branch '${branchName}' is protected!`));
+              console.log(chalk.red('This branch may be connected to your production theme.'));
+            }
+
+            if (branchInfo) {
+              console.log(chalk.yellow(`\n⚠️  Branch '${branchName}' already contains:`));
+              console.log(chalk.yellow(`- Last commit: ${branchInfo.commit.commit.message}`));
+              console.log(chalk.yellow(`- Author: ${branchInfo.commit.commit.author.name}`));
+              console.log(chalk.yellow(`- Date: ${branchInfo.commit.commit.author.date}`));
+            }
+
+            const confirmPush = await inquirer.prompt([
+              {
+                type: 'confirm',
+                name: 'proceed',
+                message: chalk.red(`Are you absolutely sure you want to push to '${branchName}'? This action cannot be undone and may affect production!`),
+                default: false
+              }
+            ]);
+
+            if (!confirmPush.proceed) {
+              console.log(chalk.blue('\nOperation cancelled. Consider creating a new branch instead.'));
+              const createNewBranch = await inquirer.prompt([
+                {
+                  type: 'confirm',
+                  name: 'create',
+                  message: 'Would you like to create a new branch instead?',
+                  default: true
+                },
+                {
+                  type: 'input',
+                  name: 'newBranchName',
+                  message: 'Enter the new branch name:',
+                  when: (answers) => answers.create,
+                  default: 'development'
+                }
+              ]);
+
+              if (createNewBranch.create) {
+                const newBranchName = createNewBranch.newBranchName;
+                execSync(`git checkout -b ${newBranchName}`, { stdio: 'inherit' });
+                execSync(`git push -u origin ${newBranchName}`, { stdio: 'inherit' });
+              } else {
+                throw new Error('Setup cancelled by user');
+              }
+            } else {
+              console.log(chalk.yellow('\n⚠️  Proceeding with push to existing branch...'));
+              // First checkout the branch locally
+              try {
+                execSync(`git checkout -b ${branchName}`, { stdio: 'inherit' });
+              } catch (error) {
+                // If branch already exists locally, just switch to it
+                execSync(`git checkout ${branchName}`, { stdio: 'inherit' });
+              }
+
+              // Then try to push, if it fails due to diverged histories, suggest pulling first
+              try {
+                execSync(`git push -u origin ${branchName}`, { stdio: 'inherit' });
+              } catch (error) {
+                console.log(chalk.yellow('\nRemote branch has different history. Attempting to sync...'));
+                try {
+                  // Fetch the remote branch first
+                  execSync(`git fetch origin ${branchName}`, { stdio: 'inherit' });
+
+                  // Try to rebase on top of remote changes
+                  try {
+                    execSync(`git rebase origin/${branchName}`, { stdio: 'inherit' });
+                  } catch (rebaseError) {
+                    // If rebase fails, abort it and try merge instead
+                    execSync('git rebase --abort', { stdio: 'inherit' });
+                    console.log(chalk.yellow('\nRebase failed, trying merge instead...'));
+                    execSync(`git merge origin/${branchName}`, { stdio: 'inherit' });
+                  }
+
+                  // Push the changes
+                  execSync(`git push -u origin ${branchName}`, { stdio: 'inherit' });
+                } catch (syncError) {
+                  console.error(chalk.red('\nFailed to sync with remote branch.'));
+                  console.error(chalk.yellow('\nTo resolve this manually:'));
+                  console.error(chalk.yellow(`1. Run: git fetch origin ${branchName}`));
+                  console.error(chalk.yellow(`2. Run: git checkout ${branchName}`));
+                  console.error(chalk.yellow(`3. Run: git merge origin/${branchName}`));
+                  console.error(chalk.yellow('4. Resolve any conflicts'));
+                  console.error(chalk.yellow('5. Run: git push origin ${branchName}'));
+                  throw new Error('Failed to sync with remote branch');
+                }
+              }
+            }
+          }
+        } else {
+          // For new repositories, default to development branch
+          execSync('git checkout -b development', { stdio: 'inherit' });
+          execSync('git push -u origin development', { stdio: 'inherit' });
+        }
+
+        // Set up secrets
+        const secrets = {
+          'SHOPIFY_STORE_URL': answers.storeUrl,
+          'SHOPIFY_CLI_THEME_TOKEN': answers.themeToken
+        };
+
+        const secretsSet = await setGitHubSecrets(repoName, secrets);
+        if (secretsSet) {
+          console.log(chalk.green('\n✨ GitHub repository and secrets set up successfully!'));
+        }
+      } else {
+        console.log(chalk.yellow('\nPlease manually add these secrets to your GitHub repository:'));
+        console.log(chalk.yellow(`- SHOPIFY_FLAG_STORE: ${answers.storeUrl}`));
+        console.log(chalk.yellow('- SHOPIFY_CLI_THEME_TOKEN: [your-theme-token]'));
+      }
+    }
+
+    console.log(chalk.green('\n✨ Setup completed successfully!'));
+    if (!answers.setupGithub) {
+      console.log(chalk.blue(`
+Next steps:
+1. Create a new GitHub repository
+2. Push your code:
+   git remote add origin https://github.com/USERNAME/REPO.git
+   git add .
+   git commit -m "feat: Initial theme setup"
+   git push -u origin main
+3. Make your first changes using conventional commits
+      `));
+    }
+  } catch (error) {
+    console.error(chalk.red('Error during setup:', error));
+  }
+}
+
 async function syncWorkflowsAndConfig(projectPath = process.cwd()) {
   console.log(chalk.blue('\nSyncing workflows and configurations...\n'));
 
   try {
+    const githubDir = path.join(projectPath, '.github');
+
+    if (fs.existsSync(githubDir)) {
+      const { shouldDelete } = await inquirer.prompt([
+        {
+          type: 'confirm',
+          name: 'shouldDelete',
+          message: 'Existing .github directory found. Would you like to delete it before syncing? (Recommended to avoid duplicates)',
+          default: true
+        }
+      ]);
+
+      if (shouldDelete) {
+        console.log(chalk.yellow('\nDeleting existing .github directory...'));
+        fs.rmSync(githubDir, { recursive: true, force: true });
+        console.log(chalk.green('✓ Deleted existing .github directory'));
+      } else {
+        console.log(chalk.yellow('\nKeeping existing .github directory. Note: This might result in duplicate or conflicting workflows.'));
+      }
+    }
+
     // Define source and target paths
-    const packagePath = path.join(__dirname, '..', 'package');
+    const templatesDir = getTemplatesDir();
+    const workflowTemplatesDir = path.join(templatesDir, 'workflows');
 
     // Files to sync
     const filesToSync = [
       {
-        source: '.github/workflows/theme-preview.yml',
+        source: path.join(workflowTemplatesDir, 'preview.yml'),
         target: '.github/workflows/theme-preview.yml',
         type: 'workflow'
       },
       {
-        source: '.github/workflows/release.yml',
+        source: path.join(workflowTemplatesDir, 'release.yml'),
         target: '.github/workflows/release.yml',
         type: 'workflow'
       },
       {
-        source: 'release.config.js',
+        source: path.join(templatesDir, 'release.config.js'),
         target: 'release.config.js',
         type: 'config'
       },
       {
-        source: 'commitlint.config.js',
+        source: path.join(templatesDir, 'commitlint.config.js'),
         target: 'commitlint.config.js',
         type: 'config'
       }
@@ -681,560 +674,44 @@ async function syncWorkflowsAndConfig(projectPath = process.cwd()) {
 
     // Sync each file
     for (const file of filesToSync) {
-      const sourcePath = path.join(packagePath, file.source);
-      const targetPath = path.join(projectPath, file.target);
-
-      if (!fs.existsSync(sourcePath)) {
+      if (!fs.existsSync(file.source)) {
         console.log(chalk.yellow(`⚠️  Source file not found: ${file.source}`));
         continue;
       }
 
+      const targetDir = path.dirname(file.target);
+      if (!fs.existsSync(targetDir)) {
+        fs.mkdirSync(targetDir, { recursive: true });
+      }
+
       // Backup existing file if it exists
-      if (fs.existsSync(targetPath)) {
-        const backupPath = `${targetPath}.backup`;
-        fs.copyFileSync(targetPath, backupPath);
+      if (fs.existsSync(file.target)) {
+        const backupPath = `${file.target}.backup`;
+        fs.copyFileSync(file.target, backupPath);
         console.log(chalk.gray(`📦 Backed up existing ${file.type}: ${file.target} → ${path.basename(backupPath)}`));
       }
 
       // Copy new file
-      fs.copyFileSync(sourcePath, targetPath);
+      fs.copyFileSync(file.source, file.target);
       console.log(chalk.green(`✓ Updated ${file.type}: ${file.target}`));
     }
 
-    console.log(chalk.green('\n✨ Successfully synced workflows and configurations!'));
-    console.log(chalk.blue('\nNext steps:'));
-    console.log('1. Review the updated files');
-    console.log('2. Check for any custom configurations in the backup files');
-    console.log('3. Commit the changes if everything looks good');
-    console.log('\nBackup files are created with .backup extension');
+    console.log(chalk.green('\n✓ Successfully synced workflows and configurations!\n'));
 
+    // Provide git instructions if changes were made
+    console.log(chalk.blue('To apply these changes:'));
+    console.log(chalk.yellow('1. Review the changes:'));
+    console.log('   git status');
+    console.log(chalk.yellow('2. Add and commit the changes:'));
+    console.log('   git add .github/ release.config.js commitlint.config.js');
+    console.log('   git commit -m "chore: Update workflows and configurations"');
+    console.log(chalk.yellow('3. Push to your repository:'));
+    console.log('   git push\n');
   } catch (error) {
-    console.error(chalk.red('\nError syncing workflows and configurations:'), error.message);
-    console.log(chalk.yellow('\nTry running the command with sudo if it\'s a permissions issue.'));
-  }
-}
-
-// New function to handle GitHub connection
-async function connectGitHub() {
-  console.log(chalk.blue('\nGitHub Repository Setup\n'));
-
-  try {
-    // Check if GitHub CLI is installed and authenticated
-    try {
-      execSync('gh auth status', { stdio: 'ignore' });
-    } catch (error) {
-      console.log(chalk.red('\n⚠️  GitHub CLI not installed or not authenticated'));
-      console.log('Please install the GitHub CLI and run gh auth login');
-      return;
-    }
-
-    // Get GitHub username
-    let username;
-    try {
-      username = execSync('gh api user -q .login', { encoding: 'utf8' }).trim();
-    } catch (error) {
-      console.error(chalk.red('Error getting GitHub username. Please ensure you are logged in with gh cli.'));
-      return;
-    }
-
-    // Get repository information
-    const answers = await inquirer.prompt([
-      {
-        type: 'list',
-        name: 'repoAction',
-        message: 'What would you like to do?',
-        choices: [
-          { name: 'Create a new repository', value: 'create' },
-          { name: 'Connect to an existing repository', value: 'connect' }
-        ]
-      },
-      {
-        type: 'input',
-        name: 'repoName',
-        message: 'Repository name:',
-        validate: input => /^[a-zA-Z0-9-_]+$/.test(input) || 'Please enter a valid repository name'
-      }
-    ]);
-
-    if (answers.repoAction === 'create') {
-      console.log(chalk.blue('\nCreating new repository...'));
-
-      try {
-        execSync(`gh repo create "${answers.repoName}" --private`, { stdio: 'inherit' });
-        console.log(chalk.green('✓ Repository created successfully'));
-      } catch (error) {
-        if (error.message.includes('already exists')) {
-          console.error(chalk.red('Repository already exists. Please choose a different name.'));
-          return;
-        }
-        throw error;
-      }
-    }
-
-    // Set up remote
-    const repoUrl = `https://github.com/${username}/${answers.repoName}`;
-
-    try {
-      execSync('git remote remove origin', { stdio: 'pipe' });
-    } catch (error) {
-      // Ignore error if remote doesn't exist
-    }
-
-    execSync(`git remote add origin ${repoUrl}.git`, { stdio: 'inherit' });
-    console.log(chalk.green('✓ Remote added successfully'));
-
-    // Set up branch protection
-    console.log(chalk.blue('\nSetting up branch protection rules...'));
-
-    try {
-      // Protect main branch
-      execSync(`gh api repos/${username}/${answers.repoName}/branches/main/protection --method PUT --field required_status_checks=null --field enforce_admins=true --field required_pull_request_reviews=null --field restrictions=null`, { stdio: 'inherit' });
-      console.log(chalk.green('✓ Branch protection rules set for main'));
-
-      // Protect staging branch
-      execSync(`gh api repos/${username}/${answers.repoName}/branches/staging/protection --method PUT --field required_status_checks=null --field enforce_admins=true --field required_pull_request_reviews=null --field restrictions=null`, { stdio: 'inherit' });
-      console.log(chalk.green('✓ Branch protection rules set for staging'));
-    } catch (error) {
-      console.log(chalk.yellow('Note: Branch protection rules will be set after first push'));
-    }
-
-    // Set up GitHub secrets
-    console.log(chalk.blue('\nSetting up GitHub secrets...'));
-
-    const secretAnswers = await inquirer.prompt([
-      {
-        type: 'password',
-        name: 'shopifyToken',
-        message: 'Enter your Shopify CLI theme token:',
-        validate: input => input.length > 0 || 'Theme token is required'
-      },
-      {
-        type: 'input',
-        name: 'shopifyStore',
-        message: 'Enter your Shopify store URL (e.g., your-store.myshopify.com):',
-        validate: input => /^[a-zA-Z0-9][a-zA-Z0-9-]*\.myshopify\.com$/.test(input) || 'Please enter a valid Shopify store URL'
-      }
-    ]);
-
-    try {
-      execSync(`gh secret set SHOPIFY_CLI_THEME_TOKEN -b"${secretAnswers.shopifyToken}" --repo=${username}/${answers.repoName}`, { stdio: 'inherit' });
-      execSync(`gh secret set SHOPIFY_STORE_URL -b"${secretAnswers.shopifyStore}" --repo=${username}/${answers.repoName}`, { stdio: 'inherit' });
-      console.log(chalk.green('✓ GitHub secrets configured successfully'));
-    } catch (error) {
-      console.error(chalk.red('Error setting GitHub secrets:'), error.message);
-      console.log(chalk.yellow('\nPlease set these secrets manually in your repository settings:'));
-      console.log('- SHOPIFY_CLI_THEME_TOKEN');
-      console.log('- SHOPIFY_STORE_URL');
-    }
-
-    console.log(chalk.green('\n✨ GitHub repository setup completed successfully!'));
-    console.log(chalk.blue('\nNext steps:'));
-    console.log('1. Push your code:', chalk.cyan('git push -u origin main'));
-    console.log('2. Create staging branch:', chalk.cyan('git checkout -b staging && git push -u origin staging'));
-    console.log('3. Review repository settings at:', chalk.cyan(repoUrl + '/settings'));
-
-  } catch (error) {
-    console.error(chalk.red('\nError during GitHub setup:'), error.message);
-  }
-}
-
-async function main() {
-  console.log(chalk.blue('Welcome to Mili Theme - Shopify Theme Automation\n'));
-
-  // Handle command line arguments
-  if (argv.help) {
-    displayCommands();
-    return;
-  }
-
-  if (argv.sync) {
-    await syncWorkflowsAndConfig();
-    return;
-  }
-
-  if (argv.listStores) {
-    await listStoredConfigs();
-    return;
-  }
-
-  if (argv.removeStore) {
-    await removeStoredConfig(argv.removeStore);
-    return;
-  }
-
-  if (argv.connectGithub) {
-    await connectGitHub();
-    return;
-  }
-
-  // Check if we're in an existing Shopify theme directory
-  const isExistingTheme = fs.existsSync('config') && fs.existsSync('layout') && fs.existsSync('templates');
-  const templatesDir = getTemplatesDir();
-  const hasTemplates = fs.existsSync(templatesDir);
-
-  // Only require templates directory if we're not in an existing theme directory
-  if (!isExistingTheme && !hasTemplates) {
-    console.error(chalk.red('Error: Not in a Shopify theme directory and no templates found. Please run this command in a Shopify theme directory or ensure the package is installed correctly.'));
-    process.exit(1);
-  }
-
-  // Ensure .gitignore is set up properly
-  ensureGitignore();
-
-  // Set up GitHub workflow files
-  setupGitHubWorkflows();
-
-  try {
-    let answers;
-
-    // Try to use stored configuration if --use-stored flag is present
-    if (argv.useStored) {
-      const storedConfig = await getStoredStoreConfig();
-      if (storedConfig) {
-        answers = {
-          projectName: storedConfig.storeName,
-          storeName: storedConfig.storeName,
-          themeToken: storedConfig.themeToken
-        };
-        console.log(chalk.green(`Using stored configuration for ${storedConfig.storeName}`));
-      } else {
-        console.log(chalk.yellow('No stored configurations found. Proceeding with setup questions.'));
-      }
-    }
-
-    // If no stored config or --use-stored not present, ask questions
-    if (!answers) {
-      answers = await inquirer.prompt([
-        {
-          type: 'input',
-          name: 'projectName',
-          message: 'What is your client/project name?',
-          validate: input => input.length > 0 || 'Project name is required'
-        },
-        {
-          type: 'input',
-          name: 'storeName',
-          message: 'What is your Shopify store name? (e.g. your-store)',
-          validate: input => /^[a-zA-Z0-9][a-zA-Z0-9-]*$/.test(input) || 'Please enter a valid store name'
-        }
-      ]);
-
-      // Get theme token (either stored or new)
-      answers.themeToken = await getThemeToken(answers.storeName, argv.useStored);
-    }
-
-    // Construct the full store URL
-    const storeUrl = `${answers.storeName}.myshopify.com`;
-
-    console.log(chalk.green('\nCreating theme directory structure...\n'));
-
-    // Create directory structure
-    const directories = ['assets', 'config', 'layout', 'locales', 'sections', 'snippets', 'templates'];
-    directories.forEach(dir => {
-      if (!fs.existsSync(dir)) {
-        fs.mkdirSync(dir, { recursive: true });
-      }
-    });
-
-    // Initialize Git repository
-    console.log(chalk.green('\nInitializing Git repository...\n'));
-
-    try {
-      if (!fs.existsSync('.git')) {
-        execSync('git init', { stdio: 'inherit' });
-        execSync('git checkout -b main', { stdio: 'inherit' });
-      }
-    } catch (error) {
-      console.error(chalk.yellow('Warning: Git initialization failed. You may need to set up Git manually.'));
-      console.error(error.message);
-    }
-
-    // Create package.json
-    console.log(chalk.green('\nCreating package.json...\n'));
-
-    const packageJson = {
-      name: answers.projectName.toLowerCase().replace(/\s+/g, '-'),
-      version: '0.0.0-development',
-      private: true,
-      description: `Shopify theme for ${answers.projectName}`,
-      scripts: {
-        "theme:dev": `shopify theme dev --store=${storeUrl}`,
-        "theme:pull": `shopify theme pull --store=${storeUrl}`,
-        "theme:push:staging": `shopify theme push --store=${storeUrl} -t "[Staging] - ${answers.projectName}"`,
-        "theme:push:production": `shopify theme push --store=${storeUrl} -t "[Production] - ${answers.projectName}"`,
-        "test": "echo \"No tests configured\"",
-        "prepare": "husky install",
-        "semantic-release": "semantic-release"
-      },
-      devDependencies: {
-        '@commitlint/cli': '^18.4.0',
-        '@commitlint/config-conventional': '^18.4.0',
-        'husky': '^8.0.0',
-        '@semantic-release/changelog': '^6.0.3',
-        '@semantic-release/commit-analyzer': '^11.1.0',
-        '@semantic-release/git': '^10.0.1',
-        '@semantic-release/github': '^9.2.5',
-        '@semantic-release/release-notes-generator': '^12.1.0',
-        'semantic-release': '^23.0.0'
-      }
-    };
-
-    fs.writeFileSync('package.json', JSON.stringify(packageJson, null, 2));
-
-    // Create release.config.js
-    console.log(chalk.green('\nCreating release.config.js...\n'));
-
-    const releaseConfig = `module.exports = {
-  branches: [
-    'main',
-    {name: 'staging', prerelease: true}
-  ],
-  plugins: [
-    ['@semantic-release/commit-analyzer', {
-      preset: 'angular',
-      releaseRules: [
-        {type: 'feat', release: 'minor'},
-        {type: 'fix', release: 'patch'},
-        {type: 'chore', release: 'patch'},
-        {type: 'docs', release: 'patch'},
-        {type: 'style', release: 'patch'},
-        {type: 'refactor', release: 'patch'},
-        {type: 'perf', release: 'patch'},
-        {type: 'test', release: 'patch'}
-      ]
-    }],
-    '@semantic-release/release-notes-generator',
-    ['@semantic-release/changelog', {
-      changelogFile: 'CHANGELOG.md'
-    }],
-    ['@semantic-release/git', {
-      assets: ['CHANGELOG.md', 'package.json'],
-      message: 'chore(release): \${nextRelease.version} [skip ci]\\n\\n\${nextRelease.notes}'
-    }],
-    ['@semantic-release/exec', {
-      prepareCmd: 'echo \${nextRelease.version} > version.txt',
-      publishCmd: 'shopify theme push -t \${nextRelease.version}'
-    }],
-    ['@semantic-release/github', {
-      assets: [
-        {path: 'version.txt', label: 'Theme Version'},
-        {path: 'CHANGELOG.md', label: 'Changelog'}
-      ]
-    }]
-  ]
-};`;
-
-    fs.writeFileSync('release.config.js', releaseConfig);
-
-    // Create commitlint.config.js
-    console.log(chalk.green('\nCreating commitlint.config.js...\n'));
-
-    const commitlintConfig = `module.exports = {
-  extends: ['@commitlint/config-conventional'],
-  rules: {
-    'type-enum': [2, 'always', [
-      'feat',
-      'fix',
-      'docs',
-      'style',
-      'refactor',
-      'perf',
-      'test',
-      'chore',
-      'revert'
-    ]],
-    'scope-case': [2, 'always', 'kebab-case'],
-    'subject-case': [2, 'always', 'sentence-case'],
-    'body-max-line-length': [2, 'always', 100]
-  }
-};`;
-
-    fs.writeFileSync('commitlint.config.js', commitlintConfig);
-
-    // Create README.md
-    console.log(chalk.green('\nCreating README.md...\n'));
-
-    const readmeContent = `<p align="center">
-  <img src="https://storage.googleapis.com/msgsndr/13wKtghjM0UiJUVzcCwH/media/678ec74139a4c6edc3269a76.png" alt="${answers.projectName} Logo" width="200"/>
-</p>
-
-<h1 align="center">${answers.projectName} Shopify Theme</h1>
-
-<p align="center">
-  <a href="#-features">Features</a> •
-  <a href="#-getting-started">Getting Started</a> •
-  <a href="#-development">Development</a> •
-  <a href="#-deployment">Deployment</a> •
-  <a href="#-contributing">Contributing</a>
-</p>
-
-## ✨ Features
-
-- 🚀 Automated semantic versioning
-- 🔄 CI/CD with GitHub Actions
-- 📦 Shopify theme development tools
-- 🛠️ Modern development workflow
-- 📝 Automated changelog generation
-- 🔐 Secure theme deployment
-
-## 🎯 Theme Structure
-
-\`\`\`
-📁 Theme Root
-├── 📁 assets/           # Theme assets (CSS, JS, images)
-├── 📁 config/          # Theme settings and schema
-├── 📁 layout/          # Theme layout templates
-├── 📁 locales/         # Translation files
-├── 📁 sections/        # Theme sections
-├── 📁 snippets/        # Reusable template snippets
-└── 📁 templates/       # Page templates
-\`\`\`
-
-## 🚀 Getting Started
-
-### Prerequisites
-
-| Tool | Version | Description |
-|------|---------|-------------|
-| Node.js | >= 16 | JavaScript runtime |
-| Shopify CLI | Latest | Shopify development tools |
-| GitHub CLI | Latest | GitHub integration |
-
-### Quick Start 🏃‍♂️
-
-1. **Install Dependencies**
-   \`\`\`bash
-   npm install
-   \`\`\`
-
-2. **Start Development**
-   \`\`\`bash
-   npm run theme:dev
-   \`\`\`
-
-## 💻 Development
-
-### Available Scripts
-
-| Command | Description |
-|---------|-------------|
-| \`npm run theme:dev\` | Start development server |
-| \`npm run theme:pull\` | Pull theme from Shopify |
-| \`npm run theme:push:staging\` | Push to staging theme |
-| \`npm run theme:push:production\` | Push to production theme |
-| \`npm run semantic-release\` | Run semantic release |
-
-## 📦 Semantic Release
-
-This project uses semantic-release to automate version management. Version numbers follow [Semantic Versioning](https://semver.org/).
-
-### 📝 Commit Message Format
-
-| Type | Description | Version Impact |
-|------|-------------|----------------|
-| \`feat\` | New feature | Minor bump |
-| \`fix\` | Bug fix | Patch bump |
-| \`docs\` | Documentation | No bump |
-| \`style\` | Code style | No bump |
-| \`refactor\` | Code refactoring | Patch bump |
-| \`perf\` | Performance | Patch bump |
-| \`test\` | Testing | No bump |
-| \`chore\` | Maintenance | Patch bump |
-| \`revert\` | Revert changes | Patch bump |
-
-### 🌳 Branch Strategy
-
-| Branch | Purpose | Protection |
-|--------|---------|------------|
-| \`main\` | Production | 🔒 Protected |
-| \`staging\` | Pre-production | 🔒 Protected |
-| \`feat/*\` | Feature development | 🔓 Open |
-| \`fix/*\` | Bug fixes | 🔓 Open |
-
-## 🚀 Deployment Workflow
-
-1. 🔨 Create feature branch
-   \`\`\`bash
-   git checkout -b feat/your-feature-name
-   \`\`\`
-
-2. 💾 Commit changes using conventional commits
-   \`\`\`bash
-   git commit -m "feat: Add new feature"
-   \`\`\`
-
-3. 🔄 Create PR to \`staging\`
-4. 👀 Review preview theme
-5. 🎯 Merge to \`staging\`
-6. 🚀 Create PR from \`staging\` to \`main\`
-
-## 🔄 CI/CD Pipeline
-
-| Stage | Trigger | Actions |
-|-------|---------|---------|
-| Validate | PR, Push | 🔍 Lint commits<br>✅ Run tests |
-| Preview | PR | 🎨 Create preview theme |
-| Release | Merge to main | 📦 Create release<br>📝 Generate changelog<br>🚀 Deploy theme |
-
-## ⚙️ Theme Configuration
-
-- 🏪 Store URL: \`${storeUrl}\`
-- 🔐 Environment: Production & Staging
-- 🌐 Theme Access: Staff accounts
-
-## 📚 Resources
-
-- [Shopify Theme Development](https://shopify.dev/themes)
-- [Semantic Release](https://semantic-release.gitbook.io/)
-- [Conventional Commits](https://www.conventionalcommits.org/)
-- [GitHub Actions](https://docs.github.com/actions)
-
-## 🤝 Contributing
-
-1. Create feature branch
-2. Commit changes
-3. Create pull request
-4. Get review & approval
-5. Merge & deploy
-
-## 📄 License
-
-This project is private and confidential.
-
----
-
-<p align="center">
-  Made with ❤️ by <a href="https://github.com/pasquinphilippe">milistack shopify agency</a>
-</p>`;
-
-    fs.writeFileSync('README.md', readmeContent);
-
-    // Set up themes
-    await setupThemes(storeUrl, answers.themeToken, answers.projectName);
-
-    // Set up GitHub with store URL
-    await setupGitHub(answers.projectName, answers.themeToken, storeUrl);
-
-    console.log(chalk.green('\nSetup completed successfully! 🎉\n'));
-    console.log(chalk.blue('Next steps:'));
-    console.log('1. Create a new repository on GitHub');
-    console.log('2. Push your code:', chalk.cyan('git remote add origin <repository-url> && git push -u origin main'));
-    console.log('3. Create a new branch for your feature:', chalk.cyan('git checkout -b feat/your-feature-name'));
-    console.log('4. Start development:', chalk.cyan('npm run theme:dev'));
-    console.log('\nWhen ready to deploy:');
-    console.log('1. Create a pull request to staging');
-    console.log('2. Review the preview theme (automatically created)');
-    console.log('3. Merge to staging to update the staging theme');
-    console.log('4. Create a pull request from staging to main for production deployment');
-
-  } catch (error) {
-    console.error(chalk.red('\nError during setup:'), error);
+    console.error(chalk.red('\n❌ Error syncing workflows and configurations:'));
+    console.error(error.message);
     process.exit(1);
   }
 }
 
-// Execute main function
-main().catch(error => {
-  console.error(chalk.red('Error:'), error.message);
-  process.exit(1);
-});
+init().catch(console.error);
