--- package/.husky/commit-msg
@@ -1,5 +1,0 @@
-#!/usr/bin/env sh
-. "$(dirname -- "$0")/_/husky.sh"
-
-npx --no -- commitlint --edit $1
-npx --no -- commitlint --edit "$1"
