--- package/.npmignore
@@ -1,10 +1,0 @@
-.git/
-.github/
-.gitignore
-.env
-.env.*
-node_modules/
-test-theme/
-.husky/
-*.log
-.DS_Store
