--- package/CHANGELOG.md
@@ -1,31 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [1.0.0] - 2024-01-22
-
-### Added
-- Initial release of Mili Release CLI
-- Interactive setup process for Shopify themes
-- Semantic versioning configuration
-- GitHub Actions workflow templates
-- Automated deployment setup
-- Git hooks for commit message validation
-- Comprehensive documentation generation
-- GitHub CLI integration for repository and secrets management
-- Safety checks for branch operations
-- Template file generation and management
-
-### Changed
-- Updated package name to @shopify-tools/mili-release
-- Improved error handling and user feedback
-- Enhanced documentation with detailed setup instructions
-
-### Fixed
-- Template file reading and writing
-- GitHub Actions workflow copying
-- Branch creation and management
-- Repository URL updates in documentation
