--- package/commitlint.config.js
@@ -1,23 +1,0 @@
-export default {
-  extends: ["@commitlint/config-conventional"],
-  rules: {
-    "type-enum": [
-      2,
-      "always",
-      [
-        "feat",
-        "fix",
-        "docs",
-        "style",
-        "refactor",
-        "perf",
-        "test",
-        "chore",
-        "revert",
-      ],
-    ],
-    "scope-case": [2, "always", "kebab-case"],
-    "subject-case": [2, "always", "sentence-case"],
-    "body-max-line-length": [2, "always", 100],
-  },
-};
