--- package/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@milistack/theme-cli",
-  "version": "0.0.0-development",
-  "description": "CLI tool for setting up Shopify theme semantic releases",
-  "private": false,
-  "type": "module",
-  "main": "bin/mili-release.js",
-  "bin": {
-    "mili-theme": "./bin/mili-release.js",
-    "mili-release": "./bin/mili-release.js"
-  },
-  "files": [
-    "bin",
-    "templates",
-    "package"
-  ],
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "scripts": {
-    "theme:dev": "shopify theme dev",
-    "theme:pull": "shopify theme pull",
-    "theme:push:staging": "shopify theme push -t \"[Staging]\"",
-    "theme:push:production": "shopify theme push -t \"[Production]\"",
-    "test": "echo \"No tests configured\"",
-    "prepare": "husky install",
-    "semantic-release": "semantic-release",
-    "sync-workflows": "mili-theme --sync"
-  },
-  "dependencies": {
-    "chalk": "^5.3.0",
-    "inquirer": "^9.2.12",
-    "yargs": "^17.7.2"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^18.4.0",
-    "@commitlint/config-conventional": "^18.4.0",
-    "husky": "^8.0.0",
-    "@semantic-release/changelog": "^6.0.3",
-    "@semantic-release/commit-analyzer": "^12.0.0",
-    "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^9.2.6",
-    "@semantic-release/npm": "^11.0.0",
-    "@semantic-release/release-notes-generator": "^13.0.0",
-    "semantic-release": "^23.0.0"
-  }
-}
