+++ release.config.js
@@ -1,15 +1,7 @@
 module.exports = {
   branches: [
-    {
-      name: 'main',
-      channel: 'latest',
-      prerelease: false
-    },
-    {
-      name: 'staging',
-      channel: 'next',
-      prerelease: true
-    }
+    'main',
+    {name: 'staging', prerelease: true, channel: 'next'}
   ],
   plugins: [
     ['@semantic-release/commit-analyzer', {
@@ -31,27 +23,14 @@ module.exports = {
     ['@semantic-release/npm', {
       npmPublish: true,
       pkgRoot: '.',
-      tarballDir: 'dist',
-      distTag: process.env.BRANCH === 'staging' ? 'next' : 'latest'
+      tarballDir: 'dist'
     }],
     ['@semantic-release/git', {
       assets: ['CHANGELOG.md', 'package.json'],
       message: 'chore(release): ${nextRelease.version} [skip ci]\n\n${nextRelease.notes}'
     }],
     ['@semantic-release/github', {
-      assets: [
-        {path: 'dist/*.tgz', label: 'NPM package'},
-        {path: 'CHANGELOG.md', label: 'Changelog'}
-      ],
-      failComment: `🚨 Release failed due to invalid branch. Please ensure you're merging into 'staging' first.`,
-      successComment: `
-🎉 This PR is included in version \${nextRelease.version}
-
-The release is available on:
-- [npm package (@\${process.env.BRANCH === 'staging' ? 'next' : 'latest'} dist-tag)](https://www.npmjs.com/package/@milistack/theme-cli/v/\${nextRelease.version})
-- [GitHub release](https://github.com/pasquinphilippe/mili-release/releases/tag/v\${nextRelease.version})
-
-Your **[semantic-release]** bot :package::rocket:`
+      assets: 'dist/*.tgz'
     }]
   ]
-};
+}
